# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: "TriCore Agent API"
  description: "三核融合智能体 REST API - 意识感(白龙马) + 执行力(龙虾) + 成长性(爱马仕)"
  version: "2.4.0"
  contact:
    name: "TriCore Team"
  license:
    name: "MIT"
servers:
  - url: "http://localhost:3721"
    description: "本地开发服务器"
  - url: "https://api.tricore.example.com"
    description: "生产服务器"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    MemorySearchRequest:
      type: object
      properties:
        query:
          type: string
        limit:
          type: integer
          default: 10
...
    RAGAddDocumentRequest:
      type: object
      required:
        - title
        - content
      properties:
        title:
          type: string
        content:
          type: string
        metadata:
          type: object
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/memory/search:
    post:
      tags: [Memory]
      summary: "搜索记忆"
      operationId: searchMemories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemorySearchRequest"
      responses:
        "200":
          description: "搜索结果"
  /api/rag/document:
    post:
      tags: [RAG]
      summary: "添加RAG文档"
      operationId: addDocument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RAGAddDocumentRequest"
      responses:
        "200":
          description: "文档已添加"
...
  /api/multimodal/analyze:
    post:
      tags: [Multimodal]
      summary: "分析图片"
      operationId: analyzeImage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                imagePath:
                  type: string
                prompt:
                  type: string
      responses:
        "200":
          description: "分析结果"
...
  /api/audit/report:
    post:
      tags: [Audit]
      summary: "生成合规报告"
      operationId: generateReport
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                startDate:
                  type: integer
                endDate:
                  type: integer
      responses:
        "200":
          description: "合规报告"
...